backend/requirements.txt
# Flask Web框架
Flask==3.0.0
Flask-CORS==4.0.0

# 配置管理
python-dotenv==1.0.0

# 数据验证
pydantic==2.5.0

# 异步任务处理（未来扩展）
celery==5.3.0
redis==5.0.1

# 现有代码依赖
openai==1.12.0
torch==2.9.1
soundfile==0.12.1
numpy==1.26.0

# 开发依赖
pytest==7.4.0
black==23.11.0
flake8==6.1.0